requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.3.0+cpu
torchvision==0.18.0+cpu
Pillow>=10.0.0,<11.0.0
transformers>=4.34.0,<4.35.0
